Update medusa.py
Updated the 'do_compile' function to add the 'setImmediate' to the hooks. In that way, when attaching methods such as the 'onCreate' which are fastest that Frida, it will try to be called on Frida’s JavaScript thread as soon as possible. This could avoid the user to think that a method is not called when hooking and tracing it.